requirements.txt
python-telegram-bot==21.6
openai==1.54.4
python-dotenv==1.0.1
